Update nlp task services and independent microservices
