+++ rest/question-answering/entrypoint.sh
