+++ rest/token-tagging/Dockerfile
