+++ rest/token-tagging/app/main.py
@@ -1,0 +1,71 @@
+import os
+import logging
+from typing import List
+
+import adaptnlp
+
+import uvicorn
+from fastapi import FastAPI
+
+from .data_models import (
+    TokenTaggingRequest,
+    TokenTaggingResponse,
+)
+
+app = FastAPI()
+
+#####################
+### Initialization###
+#####################
+
+# Initialize Logging
+logger = logging.getLogger(__name__)
+logging.basicConfig(
+    level="INFO",
+    format="%(process)d-%(levelname)s-%(message)s",
+    datefmt="%Y-%m-%d %H:%M:%S",
+)
+
+# Global Modules
+_TOKEN_TAGGER = adaptnlp.EasyTokenTagger()
+
+# Get Model Configurations From ENV VARS
+_TOKEN_TAGGING_MODE = os.environ["TOKEN_TAGGING_MODE"]
+_TOKEN_TAGGING_MODEL = os.environ["TOKEN_TAGGING_MODEL"]
+
+# Event Handling
+@app.on_event("startup")
+async def initialize_nlp_task_modules():
+    _TOKEN_TAGGER.tag_text(text="", model_name_or_path=_TOKEN_TAGGING_MODEL)
+
+
+######################
+### AdaptNLP API ###
+######################
+@app.get("/")
+async def root():
+    return {"message": "Welcome to AdaptNLP"}
+
+
+@app.post("/api/token_tagger", response_model=List[TokenTaggingResponse])
+async def token_tagger(token_tagging_request: TokenTaggingRequest):
+    text = token_tagging_request.text
+    sentences = _TOKEN_TAGGER.tag_text(
+        text=text, model_name_or_path=_TOKEN_TAGGING_MODEL
+    )
+
+    payload = [sentence.to_dict(tag_type=_TOKEN_TAGGING_MODE) for sentence in sentences]
+
+    # Need a better way to serialize
+    for p in payload:
+        entities = p["entities"]
+        for e in entities:
+            labels = e["labels"]
+            e["value"] = labels[0].to_dict()["value"]
+            e["confidence"] = labels[0].to_dict()["confidence"]
+
+    return payload
+
+
+if __name__ == "__main__":
+    uvicorn.run(app, host="0.0.0.0", port=5000)
